Made alice_vsd running
